Make restarts more robust
Here we actually swap the chat adapter. We were running into issues with
the Slack library bringing the whole application down because it doesn't
throttle or cache requests. With the frequency of restarts, it would end
up crashing due to throttling on Slack's end.

Additionally, we only restart the application if there's actually a
change to the configuration.